Increase jest timeouts (#71)
[Problem]
Heroku CI/CD is encountering transient failures due to the first test in
`SignInPage` and `SignUpPage` tests. Test itself is not buggy, it is
always the first test that takes the longest to run even on local.
Likely a heroku issue as tests run in ~300ms on local.
[Solution]
Increase jest timeout from 5 to 10 seconds for the first test in those
suites.