[CA-216] Remember me revamp (#127)
[Problem]
Currently, the 'remember me' checkbox determines the length of the jwt
token expiration. This required a process to determine if a jwt token
had expired and log the user out if so. An attempt at this was done
13844c2, with the expectation that a `useEffect` was being triggered by
the jwt token expiration and the run would receive a 401 not authorized
error from a backend call, which would then log the user out. This was
determined not to be the case, the feature success was due to constant
calls to the `/api/test` endpoint, which allowed the system to log out
right when the jwt token expired. Using varying jwt expiration times
also isn't expected behavior for a 'remember me' checkbox, not being
logged in after closing out of the website is the expected behavior.

[Solution]
- Store user uid and token in session storage (short term) or local
storage (long term) depending on remember me checkbox
- New `GetUID` service to retrieve uid from both local or session
storage
- Use `GetUID` service in `AccountDetails`, `FeedbackForm`, and
`TranslationHistory`
- Remove JWT expiration timers
- Clear both session and local storage on logout
- Modify auth check in the navbar to remove infinite backend calls
- Remove excess logging